release: v4.2.1 - security hardening and driver stability fixes
Network stack: TCP SMP race fix, SYN_RECV/ESTABLISHED congestion
control init, ARP cache locking, TCP window reflect actual buffer,
DNS bounds check, RTO backoff reset, congestion slot validation

Security: BPF indirect load overflow check, cap_fd_alloc permission
validation, BPF_RET BPF_A register fix, fd_derive refcount leak fix

Drivers: NVMe PRP alignment, queue full detection, status mask fix,
resource leak fix, CFS timeout handling, CID overflow; VirtIO mfence
barrier, timeout DMA safety, descriptor chain leak fix; PCI 64-bit
BAR; IOAPIC mask ordering; APIC timer calibration timeout

11 files changed, +262/-67 lines

+++ kernel/pci.h
@@ -91,7 +91,7 @@ struct pci_device {
     uint8_t  header_type;
     uint8_t  irq_line;
     uint8_t  irq_pin;
-    uint32_t bars[PCI_MAX_BARS];
+    uint64_t bars[PCI_MAX_BARS];     /* FIXED (v4.2.1): uint64_t for 64-bit BAR support (BUG-DRV-H8) */
     uint32_t bar_sizes[PCI_MAX_BARS];
     uint32_t subsystem_vendor;
     uint32_t subsystem_id;
